refactor(slug): move bake side behind ./bake subpath, publish FlSlugFontExtension
Restructures the slug bake side to match the codified per-format `./bake`
subpath convention (precedent: @three-flatland/normals). The packer
(packBaked) and the registerable FL_slug_font glTF-Transform extension class
(FlSlugFontExtension) now live in src/bake.ts, reachable only through the
@three-flatland/slug/bake subpath. baked.ts is now glTF-Transform-free —
runtime reader + shared types only, importing solely the zero-static-dep
@three-flatland/asset (.) entry.

This keeps @gltf-transform/core out of the browser runtime static graph:
dist/baked.js, dist/SlugFontLoader.js, and dist/index.js no longer import
@gltf-transform/core or @three-flatland/asset/bake — only dist/bake.js does.
@gltf-transform/core moves from devDependencies to dependencies (a real dep
reachable via ./bake). Registering FlSlugFontExtension on a NodeIO/WebIO lets
gltf-transform tooling read and round-trip .slug.glb without dropping the
font-data accessors.
