+++ packages/slug/src/cli.ts
@@ -26,7 +26,7 @@ import { parseFont } from './pipeline/fontParser'
 import { packTextures } from './pipeline/texturePacker'
 import { bakeStrokeForGlyph } from './pipeline/strokeOffsetter'
 import type { CapStyle, JoinStyle } from './pipeline/strokeOffsetter'
-import { packBaked } from './baked'
+import { packBaked } from './bake'
 import type { BakedJSON } from './baked'
 import { writeFile } from 'node:fs/promises'
 import type { SlugGlyphData } from './types'
