+++ packages/slug/src/pipeline/textMeasureBaked.test.ts
@@ -7,7 +7,8 @@ import { parseFont } from './fontParser'
 import { packTextures } from './texturePacker'
 import { measureText } from './textMeasure'
 import { measureTextBaked } from './textMeasureBaked'
-import { packBaked, unpackBaked } from '../baked'
+import { unpackBaked } from '../baked'
+import { packBaked } from '../bake'
 import type { BakedFontData } from '../baked'
 import type { SlugGlyphData } from '../types'
